Add GOPAC_MYIPADDRESS environment variable
This variable will be used to determine the return value of
myIpAddress() function calls. Set it to an IP address to have
that returned directly. If it does not parse as an IP address,
we will assume it is an interface name, and look up an IP address
associated with that interface instead.